[REF] Refactor JWT builder class.
This is a change to move Echo handler functions from JWT method specific
objects to JWTBuilder, a class that does not depends on JWT geneartion
methods.
